+++ tests/fixture/hooks.rb
@@ -4,6 +4,12 @@ class Hooks < TLDR
     print "\nA"
   end
 
+  def around &test
+    print "("
+    test.call
+    print ")"
+  end
+
   def test_1
     print "B"
   end
@@ -15,5 +21,17 @@ class Hooks < TLDR
   def teardown
     super
     print "C"
+  end
+end
+
+class BadHook < TLDR
+  def around &test
+    print "("
+    # wups, didn't call test.call!
+    print ")"
+  end
+
+  def test_3
+    print "🎯"
   end
 end