+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+* Add `TLDR.dont_run_these_in_parallel!` method to allow tests to indicate that they
+must be run in isolation and not concurrently with any other tests
+* Add support for `around` hooks (similar to [minitest-around](https://github.com/splattael/minitest-around))
+
 ## [0.3.0]
 
 * Add `TLDR.run_these_together!` method to allow tests that can't safely be run